+++ src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { RadarVoteComponent } from './radar-vote/radar-vote.component';
 import { ResultsComponent } from './results/results.component';
 import { IndexComponent } from './index/index.component';
 import { CreateRadarComponent } from './create-radar/create-radar.component';
-import { CompareComponent } from './compare/compare.component';
+import { SelectToCompareComponent } from './select-to-compare/select-to-compare.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/', pathMatch: 'full' },
@@ -12,7 +12,7 @@ const routes: Routes = [
   { path: 'radar/:id/vote', component: RadarVoteComponent },
   { path: 'radar/:id/results', component: ResultsComponent },
   { path: 'radar', component: CreateRadarComponent },
-  { path: 'compareRadars', component: CompareComponent }
+  { path: 'selectToCompare', component: SelectToCompareComponent }
 ];
 
 @NgModule({
