+++ src/app/app.module.ts
@@ -21,8 +21,7 @@ import { ToastrModule } from 'ngx-toastr';
 import { CreateRadarComponent } from './create-radar/create-radar.component';
 import { RadarFormComponent } from './create-radar/radar-form/radar-form.component';
 import { AxesFormComponent } from './create-radar/axes-form/axes-form.component';
-import { SelectToCompareComponent } from './compare/select-to-compare/select-to-compare.component';
-import { CompareComponent } from './compare/compare.component';
+import { SelectToCompareComponent } from './select-to-compare/select-to-compare.component';
 
 @NgModule({
   declarations: [
@@ -42,8 +41,7 @@ import { CompareComponent } from './compare/compare.component';
     CreateRadarComponent,
     RadarFormComponent,
     AxesFormComponent,
-    SelectToCompareComponent,
-    CompareComponent
+    SelectToCompareComponent
   ],
   imports: [
     BrowserModule,
