--- src/app/compare/compare.component.html
@@ -1,1 +1,0 @@
-<app-select-to-compare [(firstRadar)]="firstRadar" [(secondRadar)]="secondRadar" [radars]="radars" ></app-select-to-compare>
