+++ src/app/index/index.component.html
@@ -3,7 +3,7 @@
   <div class="card card-crear-radar">
     <div class="card-body card-body-crear-radar">
       <a class="btn btn-link btn-crear-y-comparar-radar" routerLink="/radar">Crear Radar</a>
-      <a class="btn btn-link btn-crear-y-comparar-radar" routerLink="/compareRadars">Comparar Radares</a>
+      <a class="btn btn-link btn-crear-y-comparar-radar" routerLink="/selectToCompare">Comparar Radares</a>
     </div>
   </div>
