+++ src/app/select-to-compare/select-to-compare.component.html
@@ -12,7 +12,6 @@
           class="form-control select-to-compare-select"
           name="primerRadar"
           [(ngModel)]="firstRadar"
-          (ngModelChange)="onFirstRadarChange()"
         >
           <option *ngFor="let radar of radars" [ngValue]="radar">
             {{radar.title}}
@@ -26,7 +25,6 @@
           class="form-control select-to-compare-select"
           name="segundoRadarToCompare"
           [(ngModel)]="secondRadar"
-          (ngModelChange)="onSecondRadarChange()"
         >
           <option *ngFor="let radar of radars" [ngValue]="radar">
             {{radar.title}}
@@ -36,7 +34,10 @@
     </form>
 
     <div class="select-to-compare-boton-comparar">
-      <button type="button" class="btn btn-success">
+      <button
+        type="button"
+        class="btn btn-success"
+      >
         Comparar
       </button>
     </div>
