+++ src/app/select-to-compare/select-to-compare.component.scss
